Programming by Difference
This is a way forward. His example code in the book didn't quite seem to match
up with what he was saying. In working with Andrew we took this path forward.

Some differences are

MessageForwarder takes in a domain.
AnonymousMessageForwarder takes in both a domain and a list address.

These two changes enable the tests formulated on p.95 to pass witht he new
AnonymouseMessageForwarder class.