add in a test for GetValidationPercent, which was what we were trying to get to in this example so far.
The CreditValidator class was modified just a bit so that GetValidationPercent
returns a field rather than a stubbed value. The setting of that value
is still a constant stubbed in ValidateCustomer, the point is that the
value is set when calling ValidateCustomer, which is what this new
test testAllPassed100Percent is testing.