Example of testing with the Moq mocking library. Here, a mock was created of a concrete class rather than an interface. This caused us to have to make GetComponentCount on FocusWidget virtual.
Doesn't feel like the ideal scenario, more like the tail waging the dog. There
was no requirement of the application itself that necessitated this change.
It was strictly the mocking framework.

The virtual keyword is saying to the developer that this method is overridable
in a subclass, thus promoting subclassing IMO. I don't think that sprinkling
virtual on classes to get the concrete class under test is what you want
to promote. You want to promote deeper thought on the code itself, and
evolving the design with different seams. I can see how this might be helpful
in a pinch, but maybe not something to rely on as one of your first go-tos.

But how is that much different from extracting an interface and mocking it?
One thing that comes to mind is when you extract that interface you aren't
making a change to the concrete class you are extracting it out of. Extracting
an interface gives you the opportunity to think about the Interface Segregation
Principal, for example. You are still modifying the code here to make
it more testable for sure. But you're thinking about the design of the system
more deeply I feel.