This shows what it's like to use Moq with an interface that FocusWidget implements. As you can see, the testcase is almost identical. It's just the parameter to the mocking framework changes to IFocusWidget instead of FocusWidget.
The design of WaterPane has changed in that it no longer collaborates with
a FocusWidget, but rather an IFocusWidget. This opens the class up to work
with anything that implements the IFocusWidget interface, rather than
being bound to the FocusWidget class alone.

For the WaterPan class to no longer collaborate with a FocusWidget class, but
rather an IFocusWidget class, FocusWidget needed to implement the
IFocusWidget interface.