create a TestingPermitRepository class that extends PermitRepository so that we can give it some particular methods that will hide methods on the parent class.
this allows us to decouple our test class from actually calling the database,
which is what happens in the PermitRepository class. This is utilizing the
relaxing of the Singleton class PermitRepository.

i gave the TestingPermitRepository and the PermitRepository class an Id method
just to show that the class is getting re-assigned in our test.

this commit also includes a SetTestingInstance method to make use of the
Introduce Static Setter refactoring.