Here you can actually see the the Onion Parameter that SchedulingTask is.
Imagine that if I had pushed it further, giving both the Scheduler class
and the MeetingResolver class dependencies on more objects, and then
in turn those objects depend on other objects.

If that had been the case, this line of newing up a SchedulingTaskPane
would have been longer.

I deliberately did not give Scheduler and MeetingResolver further dependencies
in the inital setup as I wanted to get under way with showing the technique
for breaking the depency of an Onion Parameter. Now I kind of wish I would
have pushed it further as it would have further driven home the Onion Parameter
case.