Here we would like to sense that Run was called on the SchedulingTask.
We are having difficulties in that SchedulingTask is the real class, not
something we have been able to mock. Moq can handle this, but it requires
we make changes to our class that I would rather not make just for the sake
of the test.

This step does not compile