Use Extract Interface (p. 362) to create an interface that SchedulingTask will satisfy, but not implement itself.
The implementation of Run is still occurring in SerialTask, just like it
was before. We have created a new seam in ISchedulingTask that allows us
to break our dependencies on the inner layers of the Onion Parameter that
SchedulingTask is. We can now test to the piece of SchedulingTask that we
are interested in, that Run is called on it.

We were able to do this because Run on SerialTask is public. The ISchedulingTask
interface merely includes in it as things that SerialTask will satisfy, and C# allows us
to do this.

You'll notice also in the Init, the inner layers of our Onion (MeetingResolver
and Scheduler) are no longer dependencies we need to be concerned with or
account for.