Ran the built in code analyzer in Xamarin on the LegacyCode.Tests project and made some adjustements.
One of the recomendations it made was to remove the use of 'private' for
instance variables, as they are private by default. On the one hand, you
should code in the language you are in and rely on it's features. On the
other hand you could say including private makes it more readable... which
I somewhat agree with today. Keeping them in for now.