This is the meat of The Case of the Aliased Parameter as laid out on p. 133.
We create a FakeOriginationPermit class that can act as a stand in for
an OriginationPermit. This allows us to send in a test class, the AlwaysValidPermit
class, which will always set the _isValid flag to true when it's Validate
is called.

Since an instance of an AlwaysValidPermit is passed into the construction
of an IndustrialFacility, when we test whether our IndustrialFacility has
permits it passes because it has called the Validate of our AlwaysValidPermit
... our Fake.

Imagine that the OriginationPermit class accesses a database or has a bunch of
depedancies that are hard to fulfill. By Subclassing it and Overriding pieces
that were needed for the piece we wanted to test, we avoid creating a bunch of
parallel Interfaces to all of the permit classes as he describes on p. 134.

So Subclass and Override Method (401) can help us to break dependancies on
parameters.