This follows the pattern the Feathers describes for the Case of the Hidden Method where a method is marked as private, we would like to test it, but we don't want to make it public.
Here we have made SetSnapRegion in the CCAImage class protected so that we
can gain access to it from a subclass by calling base.SetSnapRegion (as seen
in the TestingCCAImage class). You'll notice that the same method signature
is used in the TestingCCAImage class, but I've given it the new modifier. This
allows me to make a call to the base classes version without the need to
make additional changes to the base class.