The Case of the "Helpful" Language Feature: refactoring
This comes from pages 141 - 144.

The GetKSRStreams call off of the KSR class had a dependency on HttpFileCollection
which is a sealed class owned by Microsoft. This makes writing tests for
GetKSRStreams difficult. The suggestion he gives is to use Adapt Parameter
in the absense of the ability to use Extract Interface or Extract Implementer
(can't do these because we don't own the sealed class in question).

It turns out that HttpFileCollection has an abstract super class (NameObjectCollectionBase)
that is not sealed. This is subclassed by OurHttpFileCollection to hold
files of type HttpFileCollection, and OurHttpFileCollection then becomes
the parameter of GetKSRStreams. First tricky bit solved.

Second problem is GetKSRStreams has an internal dependancy on HttpPostedFile,
another sealed class owned by Microsoft. To overcome this hard dependency,
he uses Skin and Wrap the API to get some seperation via the HttpPostedFileWrapper
class and it's implementation of IHttpPostedFile (which we create) to
solve the internal calls that GetKSRStreams needs to FileName, ContentLength
and InputStream.

This means that in our production code, prior to calling GetKSRStreams,
we would need to iterate over the original HttpFileCollection class, putting
each file in the HttpFileCollection object into an instance of OurHttpFileCollection.