Here we extract code that uses the instance of the DetailFrame into a couple of methods. Yes, one of those methods is one line and simply returns a one line call to the backing DetailFrame.
Why would we do this?

I honestly feel that some would find this not necessary and would argue
with me if I tried to do so while pairing. And I think I would have a
hard time explaining to them the intent to a point of sufficiency that
they would see the need up front.

I think he's doing this so that there are seams in between the code that
accesses GUI specific things (things you might desire to mock for instance)
and the code that doesn't care about the GUI. It's starting to give some
seperation of concerns between hard dependancies on external libraries
and our code.

This is a subtle but very important point I think.