Perform Subclass and Override Method to gain control in a testing class over the places in the AccountDetailFrame class that we needed control over in order to test that the PerformCommand method was doing the job we thought it was and should be doing.
There is a lot of subtle stuff going on here I feel. One of the things that
I think programmers often balk at is moving things from private to
protected virtual so that a testing class can override behaviour that it
needs to get control over. I think very often is the case that we would
instead push ourselves to figure out some way to mock things with a mocking
framework rather than go down a path like this.

I recently read a post by uncle bob on when he mocks and when he does not,
and as it turns out he does so very sparingly. One of the things he points
out is mocking does add time to our tests due to the reflection that has
to occur. In his opinion it also does not flex the design to move towards
better design of the system. Here is a link to the article, I'll let it
speak for itself.

https://8thlight.com/blog/uncle-bob/2014/05/10/WhenToMock.html

You can see the seams needed to get PerformCommand under test in the
GetAccountSymbol, SetDescription and SetDisplayText methods of TestingAccountDetailFrame.

I think the point I'm trying to make is consider Subclassing like this
and inserting override possiblities into the code to begin to see different
ways that the design could and maybe should be moving. I honestly think
that some would very much not be comfortable with doing these types of changes
for testing sake, I've heard it before. But it's like he says, sometimes
your design has to get a little uglier in the short term in order to make
breakthroughs in the design... and those breakthroughs be safe ones with
test coverage.